demo-app/requirements.txt
google-cloud-monitoring==0.29.0
google-cloud-pubsub==0.35.4
google-cloud-trace==0.19.0
gunicorn==19.8.1
opencensus==0.1.5
Flask==1.0.2
